Newest Version of nextLetterWithStatistics.py
Fixed a small bug, where if you typed in a correct letter in human vs. algorithm mode, you could type it again, and it wouldn't count it as wrong. Worse, it would actually count it as correct, and if you did that however many times necessary to win (as many times as there are letters in the word), it would count the win. This has been fixed.